% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{azimuth_elevation}
\alias{azimuth_elevation}
\title{Calculate elevation angles for each azimuth}
\usage{
azimuth_elevation(radius, raster_elev, point_x, point_y, plot = FALSE)
}
\arguments{
\item{radius}{Numeric. The radius around the sampling site (in metres) relevant for shielding factor calculation.}

\item{raster_elev}{The raster (an object of type "SpatRaster", created with \link{dem_relative}) showing the elevation angle with respect to the sampling site.}

\item{point_x}{Numeric. \emph{x}-coordinate of the sampling site. Can be obtained with \link{point_xyz}.}

\item{point_y}{Numeric. \emph{y}-coordinate of the sampling site. Can be obtained with \link{point_xyz}.}

\item{plot}{Logical. Should pairs of azimuth and elevation angles be plotted? Default to FALSE.}
}
\description{
This function allows for calculating elevation angles for each azimuth (1:360).
}
\details{
This function creates points along polylines that extend from the sampling site.
The length of the polylines is equal to radius.
Polylines are created for each azimuth (1:360)
The spacing of the points corresponds to the x--y resolution of raster_elev.
The points are then converted into an object of type "SpatVect" with the aid of \link[terra:vect]{terra::vect}.
The elevation at the points is extracted from raster_elev with the aid of \link[terra:extract]{terra::extract} (method = "bilinear").
The x--y coordinates of the points with the maximum elevation at each polyline are obtained to write n ESRI shapefile to the working directory with the aid of \link[terra:writeVector]{terra::writeVector} (filetype = "ESRI Shapefile").
If plot = TRUE, raster_elev and the skyline are plotted with \link[terra:plot]{terra::plot}.
}
\examples{
TEXT.
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
