% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{azimuth_elevation_horizon}
\alias{azimuth_elevation_horizon}
\title{Obtain 360 pairs of azimuth and corresponding elevation angles for a sampling site.}
\usage{
azimuth_elevation_horizon(
  dem,
  point,
  point_x,
  point_y,
  point_z,
  radius = 10000,
  plot = FALSE
)
}
\arguments{
\item{dem}{The DEM (object of class "SpatRaster", see \link[terra:SpatRaster-class]{terra::SpatRaster} for further details).}

\item{point}{Shapefile of the sampling site (object of class "SpatVect, see \link[terra:SpatVector-class]{terra::SpatVector}).}

\item{point_x}{Numeric. The \emph{x}-coordinate of the sampling site.}

\item{point_y}{Numeric. The \emph{y}-coordinate of the sampling site.}

\item{point_z}{Numeric. The \emph{z}-coordinate of the sampling site.}

\item{radius}{Numeric. The radius around the sampling site relevant for topographic shielding calculation. Default to 10000 metres.}

\item{plot}{Logical. Create a map with the skyline?}
}
\description{
This function allows for obtaining 360 pairs of azimuth and the corresponding elevation angles with the aid of a shapefile and a digital elevation model (loaded with \link{load_geodata}).
}
\details{
For the first step, this function determines the \emph{xy}-resolution of the input-DEM with \link[terra:dimensions]{terra::res}.
This function then creates point along polylines from the sampling site to the point at the maximum distance (defined by radius) for each azimuth.
The function subsequently creates points along the polylines.
The distance between the points corresponds to the \emph{xy}-resolution of the input-DEM.
The elevation at the points is then retrieved with \link[terra:extract]{terra::extract} via bilinear interpolation.
\emph{xy}-coordinates of the points are also returned to construct the skyline around the sampling site.
This function then determines the elevation angle with respect to the sampling site for all points.
The maximum elevation angle for each azimuth is subsequently retrieved.
The skyline is constructed with the \emph{xy}-coordinates of the points with the maximum elevation angles.
The skyline (created with the aid of \link[terra:vect]{terra::vect}) is then exported to the working directory with \link[terra:writeVector]{terra::writeVector}.
A vector entitled "elevation" is added to the global environment.
This vector contains the elevation angles for all azimuths (in 1:360).
Note that the exported shapefile is an "ESRI shapefile" which can be opened with geographic information system (GIS) software, such as QGIS.
}
\examples{
point_xyz(dem, point_x = 414416.5, point_y = 5316445.8, point_z = 279, radius = 10000, plot = TRUE)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com} and Stefan Hergarten, University of Freiburg, Germany)
}
