% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{crop_dem}
\alias{crop_dem}
\title{Crop the input-DEM for shielding factor calculation}
\usage{
crop_dem(dem, point, radius, plot = FALSE)
}
\arguments{
\item{dem}{The input-DEM (an object of type "SpatRaster", created with \link[terra:rast]{terra::rast} or \link{load_geodata}).}

\item{point}{The input-shapefile (object of type "SpatVect", created with \link[terra:vect]{terra::vect} or \link{load_geodata}).}

\item{radius}{Numeric. The radius around the sampling site(s) relevant for shielding factor calculation in metres.}

\item{plot}{Logical. Should the results be plotted?}
}
\description{
This function allows for cropping the selected DEM for shielding factor calculation.
}
\details{
This function extracts the portion of the DEM which is relevant for shielding factor calculation for a single sampling site.
This depends on the selected radius around the sampling site.
For this step, \link[terra:crop]{terra::crop} is used.
Please note that 500 metres are added to radius to make sure that all all raster cells relevant for shielding factor calculation are extracted.
If plot = TRUE, a map is created with \link[terra:plot]{terra::plot}.
This map shows both the DEM and the input-point.
The cropped DEM (dem) is added to the global environment.
}
\examples{
crop_dem(dem, point, radius = 10000, plot = TRUE)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
