% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{dem_relative}
\alias{dem_relative}
\title{Elevation with respect to the sampling site}
\usage{
dem_relative(dem, point_z, plot = FALSE)
}
\arguments{
\item{dem}{The input-DEM (an object of type "SpatRaster", created with \link[terra:rast]{terra::rast} or \link{load_geodata}).}

\item{point_z}{Numeric. The \emph{z}-coordinate, i.e., the elevation of the sampling site.}

\item{plot}{Logical. Create a map with the raster using the \link[terra:plot]{terra::plot} function? Default to FALSE.}
}
\description{
This function allows for creating a raster with the elevation in metres above sea-level with respect to the \emph{z}-coordinate of the sampling site.
}
\details{
To make this calculation as fast as possible, the cropped digital elevation model should preferably used.
For this step, the \link{crop_dem} function can be used.
point_z can be obtained with \link{point_xyz}.
For the first step, the function converts the input-DEM into a dataframe with the aid of \link[terra:as.data.frame]{terra::as.data.frame}.
For the second step, the difference between the elevation of the sampling site and each cell of the DEM is calculated.
Note that if z < 0, the value is replaced by point_z.
If cells in the input-DEM contain "NA" values, an error message is returned.
An object of type "SpatRaster" (dem_rel) is added to the environment.
This raster is populated with the difference between the elevation of the sampling site and each cell of the DEM through the use of \link[terra:init]{terra::init}.
The raster (dem_rel; object of type "SpatRaster") is finally added to the global environment.
If plot = TRUE, a map is created with \link[terra:plot]{terra::plot}.
}
\examples{
dem_relative(dem = dem, point_z = 1356, plot = FALSE)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
