% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{point_xyz}
\alias{point_xyz}
\title{Get the coordinates of the sampling site}
\usage{
point_xyz(dem, point, boulder_height)
}
\arguments{
\item{dem}{The input-DEM (an object of type "SpatRaster", created with \link[terra:rast]{terra::rast} or \link{load_geodata}).}

\item{point}{The input-shapefile (object of type "SpatVect", created with \link[terra:vect]{terra::vect} or \link{load_geodata}).}

\item{boulder_height}{Numeric. The boulder height in metres.}
}
\description{
This function allows for obtaining the \emph{x}-, \emph{y}-, and \emph{z}-coordinates of a sampling site.
}
\details{
For the first step, the \emph{x}- and \emph{y}-coordinates of the sampling site are extracted with the aid of \link[terra:geometry]{terra::geom}.
For the second step, the \emph{z}-coordinate is retrieved from dem with \link[terra:extract]{terra::extract} (method = "bilinear").
The boulder height in metres is added to the \emph{z}-coordinate.
The \emph{x}-, \emph{y}-, and \emph{z}-coordinates of the sampling site (vectors) are finally added to the global environment (point_x, point_y, and point_z, respectively).
}
\examples{
point_xyz(dem, point, boulder_height = 0.3)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
