% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{raster_dist}
\alias{raster_dist}
\title{Distance from the the sampling site}
\usage{
raster_dist(dem, point_x, point_y)
}
\arguments{
\item{dem}{The input-DEM (an object of type "SpatRaster", created with \link[terra:rast]{terra::rast} or \link{load_geodata}).}

\item{point_x}{Numeric. The \emph{x}-coordinate of the sampling site.}

\item{point_y}{Numeric. The \emph{y}-coordinate of the sampling site.}
}
\description{
This function allows for generating a raster showing the distance between the sampling site and the cells of the input-DEM.
The distance is needed for the calculation of the elevation angle for each raster cell.
\link{raster_elevation} performs this task.
}
\details{
To perform this calculation as fast as possible, the cropped digital elevation model should preferably be used.
For this step, the \link{crop_dem} function can be used.
point_z can be obtained with \link{point_xyz}.
The raster (raster_distance; object of type "SpatRaster") is finally added to the global environment.
}
\examples{
raster_dist(dem = dem, point_x = 414416.5, point_y = 5316445.8)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
