% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{raster_elevation}
\alias{raster_elevation}
\title{Create a raster showing the elevation angle with respect to the sampling site}
\usage{
raster_elevation(dem_rel, raster_distance, plot = FALSE)
}
\arguments{
\item{dem_rel}{The raster (an object of type "SpatRaster", created with \link{dem_relative}) showing the elevation angle with respect to the sampling site.}

\item{raster_distance}{The raster (an object of type "SpatRaster", created with \link{raster_dist}) showing the distance between the raster cells and the sampling site.}

\item{plot}{Logical. Should the raster with the elevation angle be plotted? Default to FALSE.}
}
\description{
This function allows for generating a raster showing the elevation angle (degrees) with respect to the sampling site.
}
\details{
For the first step, the function calculates the difference in elevation for each cell of the input-DEM.
The slope of the with respect to the sampling site is then calculated for each raster cell.
Elevation angles are calculated as follows: atan(slope) * 180 / pi.
The function adds an object (raster_elev) of class "SpatRaster" to the environment.
}
\examples{
raster_elevation(dem_rel = dem_relative, raster_distance = raster_distance, plot = TRUE)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
