% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{self_shield}
\alias{self_shield}
\title{Calculate elevation angles for each azimuth (self-shielding)}
\usage{
self_shield(strike, dip, plot = FALSE)
}
\arguments{
\item{strike}{Numeric. The strike of the sampling surface (in degrees).}

\item{dip}{Numeric. The dip of the sampling surface (in degrees)}

\item{plot}{Logical. Should the results be plotted? Default to FALSE.}
}
\description{
This function calculates for every azimuth (0--360 degrees) the corresponding elevation angle in degrees.
}
\details{
This function is largely based on the MATLAB code of Greg Balco's topographic shielding calculator, available at \url{http://stoneage.ice-d.org/math/skyline/skyline.m} (last access: 16 October 2024).
It adds two vectors to the global environment: self_shielding_azimuth and self_shielding_elevation, containing the azimuth in degrees from north and the corresponding elevation angle in degrees, respectively.
}
\examples{
self_shield(strike = 20, dip = 40, plot = TRUE)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com})
}
