% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{shielding_factor}
\alias{shielding_factor}
\title{Calculate the topographic shielding factor}
\usage{
shielding_factor(point, self_shielding_elevation, elevation)
}
\arguments{
\item{point}{Shapefile of the sampling site (object of class "SpatVect, see \link[terra:SpatVector-class]{terra::SpatVector}).}

\item{self_shielding_elevation}{A vector with elevation angles for each azimuth (1:360). Can be generated with \link{self_shield}}

\item{elevation}{}
}
\description{
This function calculates the unitless topographic shielding factor for a sampling site.
}
\details{
This functions compares the elevation angles obtained with \link{self_shield} (dipping surface) and \link{azimuth_elevation} (horizon).
Following Greg Balco's topographic shielding calculator (MATLAB code available at \url{http://stoneage.ice-d.org/math/skyline/skyline.m}, last access: 16 October 2024), the larger elevation angle is chosen for shielding factor calculation.
The unitless shielding factor is calculated in the same way as in Greg Balco's topographic shielding calculator.
}
\examples{
shielding_factor(self_shielding_elevation = self_shielding_elevation, elevation  = elevation)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com} and Stefan Hergarten, University of Freiburg, Germany)
}
