% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/TopoShieldCalc.R
\name{TopoShieldFact}
\alias{TopoShieldFact}
\title{Calculate topographic shielding factors}
\usage{
TopoShieldFact(radius = 10000)
}
\arguments{
\item{radius}{Numeric. The radius around the sampling site (in metres) relevant for shielding factor calculation. Default to 10000 m.}
}
\description{
This function allows for calculating topographic shielding factors for a sampling site.
}
\details{
This function requires a shapefile (.shp) with the sampling sites and a digital elevation model (DEM; .tif)
The following columns need to be added to the attribute table and populated with data: "Name", "Strike_deg", "Dip_deg", and "BouldHt", i.e., the name of the sampling site, the strike of the sampling surface (degrees), the dip of the sampling surface (degrees), and the boulder height (metres), respectively.
The function interactively loads the required geodata and performs quality checks. See \link{load_geodata} for further details.
Subsequently, the digital elevation model (DEM) is cropped.
See \link{crop_dem} for further details.
The \emph{x}-, \emph{y}-, and \emph{z}-coordinates of the sampling sites are then retrieved.
See \link{point_xyz} for further details.
The function then creates a raster with the elevation in metres above sea-level with respect to the \emph{z}-coordinates of the sampling sites.
See \link{dem_relative} for further details.
A raster showing the distance between the sampling site and the cells of the input-DEM is subsequently generated.
See the documentation of \link{raster_dist} for further details.
For the next step, a raster showing the elevation angle (degrees) with respect to the sampling site is created.
See the documentation of \link{raster_elevation} for further details.
Elevation angles for each azimuth (1:360) are then calculated.
See the documentation of \link{azimuth_elevation} for further details.
Elevation angles for each azimuth (self-shielding) are then calculated.
See the documentation of \link{self_shield} for further details.
Finally, thhe unitless topographic shielding factor for the sampling sites is determined.
See the documentation of \link{shielding_factor} for further details.
This function adds plots (azimuth versus elevation) to the working directory (.pdf files).
It also adds the skylines (ESRI shapefiles, i.e., .shp and associated files) to the working directory via \link[terra:writeVector]{terra::writeVector}.
The function writes an open document spreadsheet (.ods) entitled "shielding.ods" to the working directory via \link[readODS:write_ods]{readODS::write_ods}.
This spreadsheet contains to columns entitled "Sampling site" and "Topographic shielding factor".
}
\examples{
TopoShieldFact(radius = 10000)
}
\author{
Felix Martin Hofmann, University of Freiburg, Germany (\email{fmhofmann9892@gmail.com} and Stefan Hergarten, University of Freiburg, Germany)
}
